fileBox/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
clerk-backend-api==3.3.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
Django==5.2.7
django-cors-headers==4.9.0
django-ratelimit==4.1.0
django-smart-ratelimit==0.8.11
django_celery_results==2.6.0
djangorestframework==3.16.1
dotenv==0.9.9
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagekitio==4.2.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.2.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14